requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
basicsr==1.4.2
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
diffusers==0.36.0
executing==2.2.1
facexlib==0.3.0
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.1
filterpy==1.4.5
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
gfpgan==1.3.8
gguf==0.17.1
gradio==6.1.0
gradio_client==2.0.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
lmdb==1.7.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
realesrgan==0.3.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruff==0.14.9
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tifffile==2025.12.12
tokenizers==0.22.1
tomlkit==0.12.0
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.4
yapf==0.43.0
zipp==3.23.0
